Combine test for DioAdapter and DioInterceptor
Since all the logic is now in the `RequestHandling` mixin, tests can be simplified and ensured that all implementors run the same test suite.
